yggdrasil: Add service for external-dns
"ExternalDNS synchronizes exposed Kubernetes Services and Ingresses with
DNS providers."[1]

We need this to ease access to development clusters, by synchronizing
Services and Ingresses into a internal DNS zone. Proxy chart PR[2].

The service can be configured from yggdrasil, which implies it can also
be configured from nidhogg. This is done so mukube-configurator can
configure it (PR[3]), as the DNS zone is provisioned together with the
cluster, so it is easier to also set the zone for external-dns at the
same time.

[1] https://github.com/kubernetes-sigs/external-dns
[2] https://github.com/distributed-technologies/helm-charts/pull/264
[3] https://github.com/distributed-technologies/mukube-configurator/pull/49