Update monitoring setup
I have been fixing some of the problems with monitoring,
this should be fixed with this update.

* Update monitoring services
* Removed label requirements from service monitors
* Removed exports

These changes are made because, while using the platform,
I saw that the dashboards build in to our Grafana instance was not build right,
this was because the dashboards was build for the exporters in the kube-prometheus-stack,
and moving our exporters out changed the labels these dashboards used.

This means that this change will deprecate the exporters chart, in favor of Kube-prometheus-stack.